pyproject.toml
[tool.poetry]
name = "h2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-aiplatform = "^1.14.0"
google-cloud-pipeline-components = "^1.0.10"
kfp = "^1.8.12"
python-dotenv = "^0.20.0"
pytz = "^2022.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
